MapFactory refactored into LevelMap.
